+++ processor/changelog.md
@@ -1,3 +1,14 @@
+## 2018-08-27 0.2.0
+- Added support to automatically balance the load od receiving messages across multiple partitions.
+- Changed the overall design of EPH.
+- Instead of attaching handlers on `eph:message` and `eph:error`, now the handlers need to be passed
+as arguments to the `start()` method on EPH.
+- Removed an optional property `autoCheckpoint` and added optional properties
+   -  `checkpointManager`
+   - `onEphError`
+   - `leaseRenewInterval`
+   - `leaseDuration`
+
 ## 2018-07-16 0.1.4
 - Added an option `autoCheckpoint: false` to not checkpoint the received messages by default.